Add support for .netconfig in dotnet-counters
The [dotnetconfig](https://dotnetconfig.org) format allows hierarchical
tool settings to be persisted and read in a standards way. By adding
this support to the counters tool, all options can now be persisted to the
`.netconfig` in the current directory to streamline subsequent runs.

This is very convenient since when doing performance investigations, it's
not unusual to perform the same collection over and over, where the only
change across runs is the process being inspected. Depending on the type
of investigation typically performed by the user, they might even be the
same settings *always* for any .NET process, so the settings could be
shared for the entire user by just saving the .netconfig to the user
profile directory instead (or some ancestor folder from where it's typically
run).

The way it's implemented on top of System.CommandLine makes it
fully non-breaking change, since providing the values via the CLI will
always override whatever is in configuration. A potential improvement
for this would be to actually make the counters list, say, be augmented
by CLI args, rather than replaced. But the current approach is safest.

.netconfig is already used by a bunch of other tools and it seems like
a good fit for this tool, and maybe others that involve many arguments
that may be relatively constant across runs (such as for `dotnet-serve`).